Don't set Specific Version = true on Xbim Nuget dependencies. Causes issues when upgrading packages
Straight extraction of the assembly resolver from the GeometryEngine wrapping class - before some refactoring to address issues where Geometry cannot be located...
